upgrade/requirements.txt
fastapi
uvicorn
requests
feedparser
openai
schedule
python-dotenv
python-docx
reportlab